trying to ward off undefinied image errors

+++ app/controllers/images_controller.rb
@@ -11,7 +11,7 @@ class ImagesController < ApplicationController
     @image.building = @building
     if @image.save(image_params)
       render json: { message: "success", fileID: @image.id, bldgID: @building.id }, :status => 200
-    else 
+    else
       #  you need to send an error header, otherwise Dropzone
       #  will not interpret the response as an error:
       render json: { error: @image.errors.full_messages.join(',')}, :status => 400
@@ -21,12 +21,13 @@ class ImagesController < ApplicationController
   # DELETE /images/1
   # DELETE /images/1.json
   def destroy
-    @image.file = nil
-    if @image.destroy
-      render json: { message: "File deleted from server" }
-    else
-      render json: { message: @image.errors.full_messages.join(',') }
-    end
+    if @image
+      @image.file = nil
+      if @image.destroy
+        render json: { message: "File deleted from server" }
+      else
+        render json: { message: @image.errors.full_messages.join(',') }
+      end
   end
 
   def sort
